Lane 8: benchmark harness (8 arms x 3 tasks) + causal routePC
train_bench.py: baseline/online/tbptt/routeA/scalarLive/routePC/
frozenPhase/frozenMag x smnist/psmnist/copy, paired streams, in-job
headroom gate. routeA = per-(layer,channel,mode) complex w in the
online backward (conj(w) on Ga/Gb blocks), meta-grad through the
one-step-lookahead exact-BPTT teacher (nested autodiff through the
custom VJP; Adam/clip ignored in the chain, cvm convention).
check_routeA_meta.py: rotation 1e-16 vs rig, w=1 bitwise, fallback
bitwise, teacher remap 4e-8, meta-grad FD 1.6e-6 — all pass.

routePC (toy gate PASS): BPTT teacher replaced by the realized online
gradient on the next batch — PC0 median R_gap 0.90 vs routeA, 4/5
seeds, BPTT_CALLS=0 audited; prediction (Simonetto beta) NOT
load-bearing. pc_signal_audit.py: h_next carries the phase signal
(cos 0.81 vs BPTT, sign 0.63). Ported as exploratory arm.

Also: tbptt_window chunked scan (forward identical, backward
truncated — verified), seq2seq head for copy, ComplexParam dtype
follows params (x64 checks), ablation_generic json fix.

+++ ssm/shared/params.py
@@ -90,4 +90,6 @@ class ComplexParam(nn.Module):
 
         re = self.param("re", real_init, self.shape)
         im = self.param("im", imag_init, self.shape)
-        return re.astype(jnp.complex64) + 1j * im.astype(jnp.complex64)
+        # dtype follows the parameters (complex64 in float32 training;
+        # complex128 under x64 — used by the finite-difference checks).
+        return re + 1j * im
